EES-5238 - breaking out shared ImportDataFunction and WriteDataFunction and renaming orchestration for creating next data set version mappings, prior to adding new orchestrations
