+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Tests/ProcessorFunctionsIntegrationTest.cs
@@ -192,11 +192,13 @@ public class ProcessorFunctionsIntegrationTestFixture : FunctionsIntegrationTest
         [
             typeof(CreateDataSetFunction),
             typeof(ProcessInitialDataSetVersionFunction),
-            typeof(CreateNextDataSetVersionFunction),
-            typeof(ProcessNextDataSetVersionFunction),
+            typeof(CreateNextDataSetVersionMappingsFunction),
+            typeof(ProcessNextDataSetVersionMappingsFunction),
             typeof(DeleteDataSetVersionFunction),
             typeof(CopyCsvFilesFunction),
             typeof(ImportMetadataFunction),
+            typeof(ImportDataFunction),
+            typeof(WriteDataFilesFunction),
             typeof(HandleProcessingFailureFunction),
             typeof(HealthCheckFunctions),
             typeof(BulkDeleteDataSetVersionsFunction),
