+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Functions/ActivityNames.cs
@@ -4,15 +4,15 @@ internal static class ActivityNames
 {
     public const string CopyCsvFiles = nameof(CopyCsvFilesFunction.CopyCsvFiles);
     public const string ImportMetadata = nameof(ImportMetadataFunction.ImportMetadata);
+    public const string ImportData = nameof(ImportDataFunction.ImportData);
+    public const string WriteDataFiles = nameof(WriteDataFilesFunction.WriteDataFiles);
     public const string HandleProcessingFailure = nameof(HandleProcessingFailureFunction.HandleProcessingFailure);
 
-    public const string ImportData = nameof(ProcessInitialDataSetVersionFunction.ImportData);
-    public const string WriteDataFiles = nameof(ProcessInitialDataSetVersionFunction.WriteDataFiles);
     public const string CompleteInitialDataSetVersionProcessing = 
         nameof(ProcessInitialDataSetVersionFunction.CompleteInitialDataSetVersionProcessing);
 
-    public const string CreateMappings = nameof(ProcessNextDataSetVersionFunction.CreateMappings);
-    public const string ApplyAutoMappings = nameof(ProcessNextDataSetVersionFunction.ApplyAutoMappings);
+    public const string CreateMappings = nameof(ProcessNextDataSetVersionMappingsFunction.CreateMappings);
+    public const string ApplyAutoMappings = nameof(ProcessNextDataSetVersionMappingsFunction.ApplyAutoMappings);
     public const string CompleteNextDataSetVersionMappingProcessing = 
-        nameof(ProcessNextDataSetVersionFunction.CompleteNextDataSetVersionMappingProcessing);
+        nameof(ProcessNextDataSetVersionMappingsFunction.CompleteNextDataSetVersionMappingProcessing);
 }
